fix(surveys): resolve CI failures for live-translation change
- Keep PostHogSurveyIntegration.swift under the 1200-line SwiftLint limit by
  moving self-contained helper types (survey state, condition matching, rating
  buckets) into a new PostHogSurveyMatching.swift, registered in the Xcode
  project.
- Use the non-availability-gated `_surveysConfig` backing in tests; the public
  `surveysConfig` accessor is iOS 15+ only and broke the simulator build.
- Add the new optional `updateSurvey(_:)` delegate method to the public API
  snapshot.

Generated-By: PostHog Code
Task-Id: 50affc3a-7798-47a6-94fe-c86b2b971dac
